pyproject.toml
[project]
name = "accubid-mcp"
version = "0.1.0"
description = "Accubid Anywhere MCP server with Trimble Identity auth"
dependencies = [
    "fastmcp>=2.13.0,<3.0.0",
    "python-dotenv>=1.0.1,<2.0.0",
    "aiohttp>=3.10.0,<4.0.0",
    "pyyaml>=6.0.0,<7.0.0",
    "trimble-id>=0.1.1,<0.2.0",
    "prometheus-client>=0.21.0,<1.0.0",
]

[project.scripts]
accubid-mcp = "src.main:run"
accubid-mcp-http = "src.main:run_http"

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0,<9.0.0",
    "pytest-asyncio>=0.24.0,<1.0.0",
    "pytest-cov>=5.0.0,<6.0.0",
    "pip-audit>=2.7.0,<3.0.0",
    "ruff>=0.6.0,<1.0.0",
    "mypy>=1.11.0,<2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = ["integration: marks tests that call live Accubid dependencies"]

[tool.coverage.run]
branch = true
source = ["src"]

[tool.coverage.report]
fail_under = 85
show_missing = true
skip_covered = true

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B"]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
pretty = true
warn_unused_ignores = true
check_untyped_defs = true
disable_error_code = ["import-untyped", "union-attr"]
